requirements.txt
django-cms>=3.5,<3.6
djangocms-admin-style>=1.2,<1.3
django-treebeard>=4.0,<5.0
djangocms-text-ckeditor>=3.2.1
djangocms-link>=1.8
djangocms-style>=1.7
djangocms-googlemap>=0.5
djangocms-snippet>=2.0
djangocms-video>=2.0
git+git://github.com/divio/djangocms-column#egg=djangocms-column
easy_thumbnails
django-filer
cmsplugin-filer>=1.1
Django==1.11.*
pytz
django-classy-tags>=0.7
html5lib>=0.999999,<0.99999999
Pillow>=3.0
django-sekizai>=0.9
six
psycopg2-binary
django-cacheops
django-redis
django-phonenumber-field
